# src/main/resources/taskapi-openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/taskapi/create":{"post":{"tags":["api-task-controller"],"operationId":"createTask","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Task"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Task"}}}}}}},"/taskapi/update-status/{id}":{"patch":{"tags":["api-task-controller"],"operationId":"updateStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStatusRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/taskapi/test":{"get":{"tags":["api-task-controller"],"operationId":"test","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/taskapi/get-task/{id}":{"get":{"tags":["api-task-controller"],"operationId":"getTaskById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Task"}}}}}}},"/taskapi/get-all-tasks":{"get":{"tags":["api-task-controller"],"operationId":"getAllTasks","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Task"}}}}}}}},"/taskapi/delete/{id}":{"delete":{"tags":["api-task-controller"],"operationId":"deleteTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"Task":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"},"dueDate":{"type":"string","format":"date-time"}}},"UpdateStatusRequest":{"type":"object","properties":{"status":{"type":"string"}}}}}}